Move `NativeImageDecoder` into a separate file, and convert it to a `class`
Given the size of the `src/core/evaluator.js` file, it cannot hurt to move some of its (image related) helper functionality into a separate file.
